Folder Changes
Project architecture changes based on Among Us tutorial
