+++ client/assets/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+    </head>
+    <body>
+        <script src="//cdn.jsdelivr.net/npm/phaser@3.0.0/dist/phaser.min.js"></script>
+        <script src="/socket.io/socket.io.js"></script>
+        <script src="../../client/src/index.js"></script>
+    </body>
+</html>
