+++ client/assets/text/how_to_play.html
