--- client/views/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-    </head>
-    <body>
-        <script src="//cdn.jsdelivr.net/npm/phaser@3.0.0/dist/phaser.min.js"></script>
-        <script src="/socket.io/socket.io.js"></script>
-        <script src="../../client/src/game.js"></script>
-    </body>
-</html>
